+++ Editor/AnimatorBuilder.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using TsiYuki.Core.Editor;
+using TsiYuki.Core.Animations.Editor;
 using UnityEditor;
 using UnityEditor.Animations;
 using UnityEngine;